// react-module/src/openapi/models/atomic-product-basic-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AtomicProductBasicDto
 */
export interface AtomicProductBasicDto {
    /**
     * 
     * @type {number}
     * @memberof AtomicProductBasicDto
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductBasicDto
     */
    barcodeFromId?: any;
    /**
     * 
     * @type {LayerShape}
     * @memberof AtomicProductBasicDto
     */
    layerShape?: any;
    /**
     * 
     * @type {AtomicProductDefinitionBasicDto}
     * @memberof AtomicProductBasicDto
     */
    atomicProductDefinition?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductBasicDto
     */
    layerStatus?: AtomicProductBasicDtoLayerStatusEnum;
    /**
     * 
     * @type {number}
     * @memberof AtomicProductBasicDto
     */
    layerNumber?: any;
    /**
     * 
     * @type {StorageAreaBasicDto}
     * @memberof AtomicProductBasicDto
     */
    storageArea?: any;
}

/**
    * @export
    * @enum {string}
    */
export enum AtomicProductBasicDtoLayerStatusEnum {
    INITIAL = 'INITIAL',
    SUCCESS = 'SUCCESS',
    BROKEN = 'BROKEN',
    ISPARTOFANOTHER = 'IS_PART_OF_ANOTHER',
    ISSEPARATED = 'IS_SEPARATED',
    INSTOCK = 'IN_STOCK',
    DOESNOTEXIST = 'DOES_NOT_EXIST'
}